+++ common/src/main/java/dev/terminalmc/claimpoints/mixin/MixinConnectScreen.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2025 TerminalMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package dev.terminalmc.claimpoints.mixin;
+
+import dev.terminalmc.claimpoints.ClaimPoints;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.screens.ConnectScreen;
+import net.minecraft.client.multiplayer.ServerData;
+import net.minecraft.client.multiplayer.TransferState;
+import net.minecraft.client.multiplayer.resolver.ServerAddress;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(ConnectScreen.class)
+public class MixinConnectScreen {
+    @Inject(
+            at = @At("HEAD"),
+            method = "connect"
+    )
+    private void connect(Minecraft client, ServerAddress address, ServerData data,
+                         TransferState transferState, CallbackInfo ci) {
+        ClaimPoints.lastConnectedIp = data.ip;
+    }
+}