+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Fixed some minor issues
 - Added Russian translation (rfin0)
+- Added option to automatically send commands such as /cp worlds on server join